helpers sweep: exports + object_link + 5 more gone; ObjectTitle + EditDestroyIcons own page-chrome; ShowPrevNextNav + IndexPaginationNav move to layouts/header
ExportsHelper: image_exporter + export_link land as
Views::Controllers::Images::Exports::Button (rendered from the
Images::ExportsController turbo_stream block). images/show.rb swaps
register_output_helper :export_status_controls /
:export_status_ml_controls for direct
render(Components::ExportStatusControls.new(...)). A throwaway
helper-vs-component parity test caught a real drift on the way out —
the component had been omitting `text-nowrap` on the not-ok bold to
mirror a quirk in the legacy helper; restored on the user's call since
the helper's inconsistency was the bug.

ObjectLinkHelper: only user_link still had a production caller
(versions/table.rb); swap to render Components::UserLink directly, then
delete the helper file + its test.

LinkHelper: delete link_with_query, icon_link_with_query, download_button,
post_button, destroy_button, edit_button, put_button, patch_button,
active_link_to, active_link_with_query, external_link. Callers in
header/context_nav_helper and header/interest_and_edit_icons_helper now
render Components::CrudButton::* directly. The remaining helpers
(modal_link_to / icon_link_to / link_icon / LINK_ICON_INDEX) stay until
autocompleter_helper.rb is converted — that is a separate PR.

LocalizationHelper: drop rank_as_lower_string + rank_as_plural_string
(zero callers).

Header::TitleHelper: extract the rich-title computation into
Views::Layouts::Header::ObjectTitle (object + user + mode: :show |
:edit). Owns the Observation ConsensusNameLink branch, the model
page_title arity dispatch, and the type-tag fallback. Deletes
page_title_for, observation_page_title, show_page_title, fallback_title,
show_title_id_badge from the helper. edit_page_title + new_page_title
stay for now (six controllers still call them as helpers.* for modal
turbo_stream renders; follow-up PR pushes that into the modal-form
component).

InterestAndEditIconsHelper: add_edit_icons shrinks to a content_for
shim around a new Views::Layouts::Header::EditDestroyIcons. The
permission gating (can_edit_object?, can_destroy_object?,
can_destroy_location?) moves with it onto the component as private
methods.

ShowPrevNextNav + IndexPaginationNav: move out of Components and into
Views::Layouts::Header (page chrome, not reusable UI primitives). The
IndexPaginationNav rework drops the opaque `args:` Hash for a typed
`anchor:` prop (the only non-empty caller passed `{ anchor: ... }`); the
internal page-link params Hash that was being mutated through `@args`
becomes a regular instance var. q_params is sourced inside the view
via `q_param(current_query)` rather than passed in from the helper —
tests stub `current_query` with a saved Query when they need to verify
the hidden q[...] fields.

UsersController: the now-deleted UserStatsHelper logic moves into
UsersController::UserStatsBuilder concern. The Phlex view UserStats
no longer takes a `paths:` prop — a FIELD_PATHS lookup inside the view
builds each link via Rails route helpers directly.

Test trims: link_helper_test + object_link_helper_test delete (no
methods left to test). articles_controller_test's
helpers.destroy_button assertion updates to
render(Components::CrudButton::Delete.new(...)).
