+++ app/helpers/link_helper.rb
@@ -1,49 +1,6 @@
 # frozen_string_literal: true
 
-#  link_with_query              # link_to with query params
-#  destroy_button               # button to destroy object
-#  post_button                  # button to post to a path
-#
-#  TO USE CAPTURE &BLOCK
-#  content = block_given? ? capture(&block) : name
-#  probably need content.html_safe.
-#  https://stackoverflow.com/questions/1047861/how-do-i-create-a-helper-with-block
-#  heads up about button_to input vs button
-#  https://blog.saeloun.com/2021/08/24/rails-7-button-to-rendering
-
 module LinkHelper
-  # Call `link_to` with query params added.
-  # Should now take exactly the same args as `link_to`.
-  # You can pass a hash to `path`, but not separate args. Can take a block.
-  def link_with_query(text = nil, path = nil, **opts, &block)
-    link = block ? text : path # first two positional, if block then path first
-    content = block ? capture(&block) : text
-
-    link_to(add_q_param(link), opts) { content }
-  end
-
-  # https://stackoverflow.com/questions/18642001/add-an-active-class-to-all-active-links-in-rails
-  # https://stackoverflow.com/questions/75742517/how-to-highlight-active-nav-link-when-using-hotwire
-  # Make a link that is a target for the stimulus "nav-active_controller"
-  # (The controller adds .active class if it's a link to the current page,
-  # and updates the active link when navigating. Allows nav to be cached!)
-  # Delegates to `Components::ActiveLink` — render the component
-  # directly in Phlex views.
-  def active_link_to(text = nil, path = nil, **, &block)
-    link_path = block ? text : path # positional: block ⇒ first arg is path
-    content = block ? capture(&block) : text
-
-    render(Components::ActiveLink.new(content, link_path, **))
-  end
-
-  # mixes in "active" class
-  def active_link_with_query(text = nil, path = nil, **, &block)
-    link = block ? text : path # because positional
-    content = block ? capture(&block) : text
-
-    active_link_to(add_q_param(link), **) { content }
-  end
-
   # Link should be to a controller action that renders the form in the modal.
   # Stimulus modal-toggle controller fetches the form from the link as a
   # turbo-stream response. It also checks if it needs to generate a modal, or
@@ -74,17 +31,6 @@ module LinkHelper
     render(Components::IconLink.new(content, link_path, **opts))
   end
 
-  # NOTE: above re: MO tabs
-  def icon_link_with_query(text = nil, path = nil, options = {}, &block)
-    return unless text
-
-    link = block ? text : path # because positional
-    content = block ? capture(&block) : text
-    opts = block ? path : options
-
-    icon_link_to(add_q_param(link), opts) { content }
-  end
-
   # Glyphicon `<span>` with the MO `link-icon` class. Pass `title:`
   # for a tooltip + screen-reader label. Delegates to
   # `Components::LinkIcon` — render the component directly in Phlex
@@ -99,22 +45,6 @@ module LinkHelper
              data: args[:data] || {},
              attributes: args.except(:title, :class, :data)
            ))
-  end
-
-  # Renders an `<a>` for an `ExternalLink` AR record. Delegates to
-  # `Components::ExternalSiteLink` — render the component directly
-  # in Phlex views. (Component is named for the destination — an
-  # external site — to avoid collision with the AR model.)
-  #
-  # `concat` preserves the long-standing side-effect emission
-  # contract of this helper, so ERB callers can write a bare
-  # `external_link(link)` line inside a `tag.li(...) do ... end`
-  # block (no `<%= %>` / `concat` wrapping needed at the call
-  # site). The previous ERB-only helper used `concat` for this
-  # reason; the Phlex-component refactor (PR #4401) accidentally
-  # dropped the `concat` and the lone caller fell silent.
-  def external_link(link)
-    concat(render(Components::ExternalSiteLink.new(link: link)))
   end
 
   # NOTE: Specific to glyphicons
@@ -179,69 +109,4 @@ module LinkHelper
     menu: "align-justify",
     info: "question-sign"
   }.freeze
-
-  # button to destroy object
-  # Used instead of link_to because method: :delete requires jquery_ujs library
-  # Sample usage:
-  #   destroy_button(target: article)
-  #   destroy_button(name: :destroy_object.t(type: :glossary_term),
-  #                  target: term)
-  #   destroy_button(
-  #     name: :destroy_object.t(type: :herbarium),
-  #     target: herbarium_path(@herbarium,
-  #     back: herbaria_path(@herbarium.try(&:id)))
-  #   )
-  #
-  def destroy_button(target:, name: nil, **args)
-    render(Components::CrudButton::Delete.new(
-             target: target, name: name, **args
-           ))
-  end
-
-  # GET-style edit link — emits `<a>` (link_to), not a form-wrapped
-  # button. Delegates to `Components::CrudButton::Edit`, which carries
-  # the `action: :edit`/`icon: :edit` defaults and (for model targets
-  # rendered from SHOW_OBS_EDITABLES controllers) the `?back=show|index`
-  # query param. Callers wanting a text-only edit link pass `icon: nil`.
-  def edit_button(target:, name: nil, **args)
-    render(Components::CrudButton::Edit.new(
-             target: target, name: name, **args
-           ))
-  end
-
-  # GET-style download link for a species_list — emits `<a>` via
-  # `Components::CrudButton::Download`. The path is built explicitly
-  # as a String because the route shape doesn't match
-  # `download_<resource>_path` — there's a `new_download_species_list`
-  # named route, but `download_species_list` is the index path.
-  def download_button(target:, name: nil, **args)
-    render(Components::CrudButton::Download.new(
-             target: new_download_species_list_path(id: target.id),
-             name: name,
-             **args
-           ))
-  end
-
-  # Refactor to accept a tab array
-
-  # POST to a path; used instead of a link because POST link requires js
-  def post_button(name:, path:, **args, &block)
-    render(Components::CrudButton::Post.new(
-             name: name, target: path, **args, &block
-           ))
-  end
-
-  # PUT to a path; used instead of a link because PUT link requires js
-  def put_button(name:, path:, **args, &block)
-    render(Components::CrudButton::Put.new(
-             name: name, target: path, **args, &block
-           ))
-  end
-
-  # PATCH to a path; used instead of a link because PATCH link requires js
-  def patch_button(name:, path:, **args, &block)
-    render(Components::CrudButton::Patch.new(
-             name: name, target: path, **args, &block
-           ))
-  end
 end
