+++ app/helpers/localization_helper.rb
@@ -9,25 +9,9 @@ module LocalizationHelper
     :"RANK_#{rank.to_s.upcase}".l
   end
 
-  # Translate Name rank (singular).
-  #
-  #   rank_as_lower_string(:genus)  -->  "genus"
-  #
-  def rank_as_lower_string(rank)
-    :"rank_#{rank.to_s.downcase}".l
-  end
-
   # Translate Name rank (plural).
   #
-  #   rank_as_plural_string(:genus)  -->  "Genera"
-  #
-  def rank_as_plural_string(rank)
-    :"RANK_PLURAL_#{rank.to_s.upcase}".l
-  end
-
-  # Translate Name rank (plural).
-  #
-  #   rank_as_plural_string(:genus)  -->  "genera"
+  #   rank_as_lower_plural_string(:genus)  -->  "genera"
   #
   def rank_as_lower_plural_string(rank)
     :"rank_plural_#{rank.to_s.downcase}".l
