+++ app/views/controllers/sequences/form.rb
@@ -141,5 +141,14 @@ module Views::Controllers::Sequences
     def observation
       @observation || model.observation
     end
+
+    # Inlined from the now-deleted `SequencesHelper#sequence_archive_options`.
+    # Dropdown options for the archive select: each entry is a
+    # `[label, value]` pair where label == value == archive name.
+    def sequence_archive_options
+      ::WebSequenceArchive.archives.each_with_object([]) do |archive, array|
+        array << Array.new(2, archive[:name])
+      end
+    end
   end
 end
