--- test/helpers/locations_helper_test.rb
@@ -1,71 +1,0 @@
-# frozen_string_literal: true
-
-require("test_helper")
-
-# Tests for LocationsHelper. The integration / system tests exercise
-# the link / count helpers transitively. Sort-options tests moved
-# to `test/controllers/locations_controller_test.rb` after the
-# `locations_index_sorts` helper was hoisted onto the controller
-# as `LocationsController#index_sort_options`.
-class LocationsHelperTest < ActionView::TestCase
-  # `find_species_list` is a defensive parser walking an arbitrary
-  # query.params hash. Each `return nil unless ...` guard below
-  # carves out a malformed-shape we don't want to crash on; these
-  # tests pin each one explicitly.
-
-  def test_find_species_list_with_object_lacking_params_method
-    assert_nil(find_species_list(Object.new))
-  end
-
-  def test_find_species_list_with_non_hash_params
-    fake_query = Struct.new(:params).new("not-a-hash")
-    assert_nil(find_species_list(fake_query))
-  end
-
-  def test_find_species_list_with_missing_observation_query_subhash
-    fake_query = Struct.new(:params).new({})
-    assert_nil(find_species_list(fake_query))
-  end
-
-  def test_find_species_list_with_non_hash_observation_query
-    fake_query = Struct.new(:params).new({ observation_query: "not-a-hash" })
-    assert_nil(find_species_list(fake_query))
-  end
-
-  def test_find_species_list_with_non_array_species_lists
-    fake_query = Struct.new(:params).new(
-      { observation_query: { species_lists: "not-an-array" } }
-    )
-    assert_nil(find_species_list(fake_query))
-  end
-
-  def test_find_species_list_with_zero_species_lists
-    fake_query = Struct.new(:params).new(
-      { observation_query: { species_lists: [] } }
-    )
-    assert_nil(find_species_list(fake_query))
-  end
-
-  def test_find_species_list_with_multiple_species_lists
-    fake_query = Struct.new(:params).new(
-      { observation_query: { species_lists: [1, 2] } }
-    )
-    assert_nil(find_species_list(fake_query))
-  end
-
-  def test_show_obs_link_title_with_count
-    loc = locations(:obs_default_location)
-    expected = "#{:show_location_observations.t} (#{loc.observations.size})"
-
-    assert_equal(expected, show_obs_link_title_with_count(loc),
-                 "Expected observation count in parentheses after label")
-  end
-
-  def test_find_species_list_with_single_species_list_returns_it
-    spl = species_lists(:first_species_list)
-    fake_query = Struct.new(:params).new(
-      { observation_query: { species_lists: [spl.id] } }
-    )
-    assert_equal(spl, find_species_list(fake_query))
-  end
-end
